RegionSelector shared component
